Moved some qgen_resp code to qgen_cube.cpp
Trying to revive the cube generation code, and related things.
Some part was moved to a special file, involved with writing and
reading cube files. More work is needed.

Fixed the output of ESP correlation xvg files in gentop, even when
ACM is used. That is you can now run with e.g. ACM-pg and get the
correlation of the potential due to generated charges with the QM
result if you have input that to the program.

Change-Id: Ia45881b5327b72d24cb088cace21226526d7eb28