Fixed serious bug in optpram.cpp stopping optimization.
For each trial parameter in optparam a copy was made before doing
a random modification and MC evaluation. In order to do the
calcDeviation, the changed parameter has to changed in poldata too.
If the change was not accepted, the internal variable was changed back,
but not poldata. This had the effect that bad parameters accumulated in
poldata and optimization would cease to work sooner or later.