Fix in generating linear angles and impropers.
Previously interactions that had the wrong order of atoms would
fail with errors generating the topology. E.g. H-C#N would work,
but not N#C-H and the same for impropers. This has been fixed by
reversing the order of atoms.

In another fix, the constant a for linear angles is now computed
in bastat.