Cleaned GROMACS related things to minimize installation.
In order not to have conflicting names and files lying around
when people also have a GROMACS installation, most things
GROMACS-related are removed here and e.g. the GMXRC is
renamed to ACTRC.

Testing of GROMACS code has been removed as well, which is
a bit dangerous since ACT relies on that code still.