Implemented calculation of dissociation energy with bootstrapping.
The code will run a user specified number of bootstraps Nb over all
the molecules. At each instance, N out of N molecules will be
randomly picked, which means some molecules will not be picked
and others multiple times. The statistics of these Nb bootstraps
will be used to compute the standard deviation of the value.

While at it, simplied the regression code to not have unnecessary
data.

TODO: make sure that no issues occur with linear algebra routines
due to linear dependence of compounds or bonds.

Part of #41
Part of #3