Change the notation for identifiers to have bond orders.
The change of identifiers has effects throughout the code
since they are used in many places.

Rewrote bastat by putting all the real work in a new file
allbondeds.cpp with header file. There is a new small test
of some of the code in that file.

Fixed fatal error message to point to ACT.

Moved some scripts to run ACT to the repository.

TODO
More checks for correctness.
More tests.

Part of #57
Part of #55