Reorganized ga code and added draft for a test.
Made the ga code independent of alexandria directory again
by reorganizing the code. The GeneticAlgorithm is now an abstract
class with two children, HybridGA and PureGA each with their own
evolve routine. The main alexandria program tune_eem decides based
on the parameters passed which one of the two to initiate.

Added a ga-test that is built automatically. It is not entirely functional
yet although it compiles and runs.

Deleted a couple outdated files in ga.