Worked on three-level parallellization.
Many changes to facilitate a three-level parallellization structure
with one overlord, a number of middlemen/individuals, each with a
number of helpers. GA works in parallel now.

New files have been added to make the code more clear, in particular
actmiddleman.cpp and acthelper.cpp.

To organize the GA code better, separate classes was created for a
Genome and for a GenePool. The Dataset was also introduced to merge
iMolSelect with the Target.

TODO: more checks to see whether MCMC and HYBRID work.
TODO: fix the ga-test.