Fixed nasty deadlock reading molecules.
It turns out that MPI library != MPI library.
There was a bug in the code with the master sending a "done" signal
to itself, but nobody was listening. This apparently is silently
ignored in some MPI libraries, like on the Apple, but not on
OpenMPI on tetralith. Not sending to itself solved the problem.