Adding a new terminator (1/2)
Now the user has the option to use the -maxTestGenerations flag, which will halt evolution of GA/HYBRID if the test fitness has not improved in that amount of generations.
This does not replace -maxGenerations, as both options can be on at the same time.
Default run does not change, as the new flag is disabled by default.