Added support for minimizing structures.
The tune_ff code will now perform an energy minimization of
each molecule after the optimization is finished. The reason
for this is that it allows us to check whether the optimized
structure from QM/DFT is similar to the one in the force field.
When optimizing for the energy (-fc_epot 1) it is prudent to also
optimize the square forces (e.g. -fc_force 0.1) to make sure the
force field gives the correct energy *and* the correct structure.

During debugging it became necessary to verify the force calculations
in the old GROMACS part of the code, for this reason a couple of
tests have been added.

The output to the log file has been updated as well with energies
before and after minimization, and a notion of the change in
structure (RMSD in pm).

TODO:
- Do a true superposition of the minimized structure on the
original before computing the RMSD.
- Make sure the optimization only happens for QM/DFT optimized
structures (JobType::OPT).