Added suport for generating charges for fragments of a molecule.
This allows the user to put in a dimer of compounds and point out
which atoms belong to which. Then the charge generation code will
use the total charge of each of the compounds and there will be
no charge transfer, or polarization between molecules.

Updated Python scripts to support the <fragments> feature.

Introduced new test for this.