Removed -lot command line options from alexandria modules.
Introduced a new function to find an experiment based on the
JobType. The mymol::GenerateAtoms now will look for a JobType::Opt
calculation only. That means all molprop.xml files need to supply
one and only one Opt calculation per compound (otherwise the results
will be undefined, this is not checked for yet).