Implemented generation of a pairlist and combining parameters.
To use the force computer for Van der Waals and Coulomb, the
parameters have to be combined. This is now implemented in the
way that the Poldata force field structure gets entries for
pairs of atom types. This is generated on the fly after reading
the input force field file. In tune_ff the routine
generateNonbondedParameterPairs will have to be called once every
time a Van der Waals parameter changes.

Part of #60.