Implemented communication by broadcast for poldata and molprop.
The goal of this exercise is to speed up initiation of FF
optimizations using tune_ff. The code can be checked using
alexandria edit -bcast [ flags ] for poldata
and
alexandria molprop_test -bcast [ flags ] for molprop.

In both cases the program needs to be run in parallel.

TODO: Implement usage in tune_ff.