Made it possible to use charges from molprop file in programs.
Reading charges from molprop files has become more general.
Three programs, gentop, tune_ff and simulate do now have an
option -qqm to read charges of a desired type from the molprop
file, if present. E.g. -qqm CM5. If this option is selected no
charges will be generated.

This is necessary in case we have multiple copies of a molecule
and we want to ascertain the charges are identical. ACM Charges can
be store in a molprop file using the alexandria molprop_check command.