Implemented code to deduce compounds/fragments from bonds.
The code relies completely on OpenBabel to generate correct bonds, but
if so, fragments or compounds are deduced from that. As a result,
individual compounds will keep their net charge. However, OpenBabel is
not 100% reliable as to charges of the compounds. Those can be set in
code but there is no user control over this as of yet.

Part of #131