Update to make ACT recognize different molecules in input files.
The code will send an input file to openbabel to generate atom types
and bonds. Then it will split the input into separate chemical moieties
and generate an InChi for each chemical moiety. This is used to distinguish
the same compound occurring multiple times in an input. Water springs to mind.

Also made the OpenMM xml generation aware of this and output distinct
"residues" corresponding to one compound each. This has the disadvantage
that compounds consisting of muliple residues, such as proteins, will be
treated as a single residue. This will have to be dealt with later.

Part of #26