Made fixes for diatomic compounds.
With the option -diatoms to tune_ff one can now print the
potential of all diatomic compounds as a function of distance.
That is, taking the input structures from the molprop file, the
reference energy minus the Coulomb energy is stored in an xvg file.

Reorganized edit_ff such that it is possible to change the parameters
for just one interaction at a time.