Made charge symmetrization a property of the force field.
This means if the force field was derived with charge symmetrization
this cannot be changed in hindsight. As a result, options to turn on
or off charge symmetrization have been moved to the gen_ff program only.

One RESP test was removed due to it being identical to another one.

Part of #162