Fix for reproducability of normal mode analysis.
Sometimes when doing a NMA two frequencies are very similar. Due to
numeric issues their order may be swapped when computed on different
hardware. This patch checks whether frequencies are very close, and if
so orders them according to the size of the intensity (largest first).

Also removed old comments from test program.

Part of #144