Fixed issue with different charges for the same compound.
In training of intermolecular as well as intramolecular energies,
the charges on identical compounds are read from a user supplied
molprop file specified with the -charges option. Once this file
is read and processed in the fragmenthandles, charge generation will
not be redone.

TODO: Implement charge map reading in other programs than gentop and train_ff

Part of #166.