Fixed major bug in handling multiple polarizabilities/energies
The code was mixing up coordinates from one molprop-experiment with
properties such as multipoles, polarizabilities as well as interaction
energies from other molprop-experiments. ESP were not affected. As a
result, fitting to those properties may have been meaningless.

Also fixed unnecessary extra printing of electostatic properties (in
fact the square of the number of e.g. dipoles was printed).

Also fixed determining of RESP charges for polarizable systems.
Even though it is not used a lot, but it was incorrect since it
used the shell force RMS that is returned by the force computer
as the convergence criterion, instead of the charges themselves.

Fixes #304