Fixed a longstanding problem in energy minimization leading to tests failing.
When minimizing the energy, we used to let ACT do the shell positions and
something else the atoms. But this may lead to minimization getting stuck
going back and forward between different atom/shell positions. Now the
forcecomputer takes an optional input flag (default true) that tells it
whether or not to minimize the shell. When the minimizer calls the compute
function, it turns off this flag. This will hopefully lead to reproducible
minimization, and code dependent on it such as normal mode analysis.

Part of #344
Fixes #153