Fixed asymmetric charges.
Despite functionality for symmetrizing charges, the code was called in the
wrong place. This led to both ACT and OpenMM using asymmetric charges in
case for instance the geometry of the molecule was not perfect.

As can be seen from all the test results that needed to be updated this was
a bug affecting a lot of code.

Fixes 564