Made charge symmetrization work at last.
Due to a subtlety in fragmenthandler, charges were symmetrized,
but upon request (fetchCharges) the unsymmetrized charges were
handed out. As a result we did not ever use symmetrized charges
and this as long as we have been using the fragmenthandler code.

Unfortunately, all the tests depending on charges failed and new
test results were needed to incorporate.

Part of #590