Made many updates to incorporate message handler without shortcuts.
The code now sports two gmx::TextWriter objects, one for normal
output, run on the master processor only, and one for debugging
on all processors. Most alexandria programs now use this for
writing to log files and the levels of logging are the same in all
programs.

TODO: Remove more "normal" printing in favor of the message handler.
TODO: Remove throws and fatal_error calls (gromacs version).

Fixes #690
Part of #215