Make missing parameters an error but only exit when needed.
By making missing parameters an error we do not allow calculations
to proceed with an incomplete force field. In practice this means
topologies do not get created without support in the force field.
It turns out that a lot of code just happily would use incorrect
topology, now crashed when there was no topology. Therefore, this patch
touches some of those routines dealing with topologies and also some
broken input files in the tests.

Fixes #808