Issue478a (#843)
* Made it possible to reuse the same parameter name

This works now, it is posible to use a parameter name,
for instance beta, in multiple interactions if it is
prepended by the interaction type in uppercase. E.g.:
train_ff -fit 'MORSE:beta INDUCTIONCORRECTION:beta'

Fixes #478

* Further work to make parameters with the same name work.

Part of #478

* Fixed missing summation to epot.

Part of #478

* Fixed missing summation of induction terms.

Probably an issue that has been around for a while without
being notices: when not specifying the -fc_deltahf flag, its
weight should be zero, and the inductioncorrection energy should
be added to the total induction. This was not done correctly.

Part of #478

---------

Co-authored-by: David van der Spoel <davidvanderspoel@gmail.com>
Co-authored-by: Najla Hosseini <62064335+najla23@users.noreply.github.com>