Implemented low level tests of all potentials. (#931)
* Implemented low level tests of all potentials.

These test compute energy and forces and store them.
For the pair potentials (including bonds) a numeric
calculation of the force is done and compared to the
analytical force, to validate the implementation of the
analytical forces.

Also fixed SEGV in actmol.cpp, calculation of interaction
energies.

Part of #928

* Fixed tests that were broken in the last patch.

Part of #928